drop the use of afarmer and merge into farmer
