+++ examples/farmer/farmer.py
@@ -17,12 +17,14 @@ import pyomo.environ as pyo
 import numpy as np
 import mpisppy.scenario_tree as scenario_tree
 import mpisppy.utils.sputils as sputils
+from mpisppy.utils import config
 
 # Use this random stream:
 farmerstream = np.random.RandomState()
 
 def scenario_creator(
     scenario_name, use_integer=False, sense=pyo.minimize, crops_multiplier=1,
+    num_scens=None
 ):
     """ Create a scenario for the (scalable) farmer example.
     
@@ -37,6 +39,9 @@ def scenario_creator(
         crops_multiplier (int, optional):
             Factor to control scaling. There will be three times this many
             crops. Default is 1.
+        num_scens (int, optional):
+            Number of scenarios. We use it to compute _mpisppy_probability. 
+            Default is None.
     """
     # scenario_name has the form <str><int> e.g. scen12, foobar7
     # The digits are scraped off the right of scenario_name using regex then
@@ -78,6 +83,10 @@ def scenario_creator(
             scen_model=model,
         )
     ]
+    
+    #Add the probability of the scenario
+    if num_scens is not None :
+        model._mpisppy_probability = 1/num_scens
     return model
 
 def pysp_instance_creation_callback(
@@ -220,6 +229,42 @@ def pysp_instance_creation_callback(
                                                sense=sense)
 
     return model
+
+# begin functions not needed by farmer_cylinders
+# (but needed by special codes such as confidence intervals)
+#=========
+def scenario_names_creator(num_scens,start=None):
+    # (only for Amalgamator): return the full list of num_scens scenario names
+    # if start!=None, the list starts with the 'start' labeled scenario
+    if (start is None) :
+        start=0
+    return [f"scen{i}" for i in range(start,start+num_scens)]
+        
+
+
+#=========
+def inparser_adder():
+    # add options unique to farmer
+    config.add_to_config("crops_multiplier",
+                         description="number of crops will be three times this (default 1)",
+                         domain=int,
+                         default=1)
+    
+    config.add_to_config("farmer_with_integers",
+                         description="make the version that has integers (default False)",
+                         domain=bool,
+                         default=False)
+
+
+#=========
+def kw_creator(options):
+    # (only for Amalgamator): linked to the scenario_creator and inparser_adder
+    kwargs = {"use_integer": options.get('use_integer', False),
+              "crops_multiplier": options.get('crops_multiplier', 1),
+              "num_scens" : options.get('num_scens', None),
+              }
+    return kwargs
+# end functions not needed by farmer_cylinders
 
 
 #============================
